Add files via upload
It is an Android based Attendance Manager app which can be used by students to manage their ever-short attendance. It provides with the ability to feed your time-table and mark your attendance on daily basis.On a given day a student can mark PRESENT, ABSENT, CANCEL or BUNK for a class. It also predicts the number of classes needed to be attended in order to make the attendance above the set criteria. Attendance for extra classes can also be marked be marked on a given day. Attendance for any previous day can also be marked.